requirements.txt
flask==3.0.3
pytesseract==0.3.10
pillow==10.3.0
opencv-python==4.9.0.80
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
jupyter==1.0.0
transformers==4.40.2
datasets==2.19.1
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
pdf2image==1.17.0
pymupdf==1.24.4
roboflow==1.1.27
python-dotenv==1.0.1
accelerate==0.29.3
seqeval==1.2.2
label-studio==1.13.1